+++ ChangeLog.md
@@ -1,8 +1,11 @@
 # ChangeLog #
 
-## Build 3.2.4.6 (04-09-2019) ##
+## Build 3.2.4.7 (04-11-2019) ##
 
+- Updated the custom default app association file.
+- The custom default association file now also gets imported into Windows 10 Enterprise LTSC 2019.
+- Updated advertising and content delivery registry settings.
+- The Immersive Control Panel page for Windows Defender is now hidden if its System Application was removed.
 - Updated the Windows Store application packages and its dependency packages.
 - Updated how objects are assigned and returned.
-- The OEM default app associations are now exported and saved before the custom default app associations are imported.
 - Additional code changes and improvements.